Done BT00

+++ 18120468_BT00/18120468_BT00/function.h
@@ -4,7 +4,7 @@
 using namespace std;
 using namespace cv;
 
-int changeContract(Mat& srcImage, Mat& desImage, short contrastValue);
+int changeContract(Mat& srcImage, Mat& desImage, float contrastValue);
 int changeBrighness(Mat& srcImage, Mat& desImage, short brighnessValue);
 int RGB2Gray(Mat& srcImage, Mat& desImage);
 int Gray2RGB(Mat& srcImage, Mat& desImage);
